e2e-tests reward points: moving functions from here to there
